[FEAT] Add italian dictionary file
