+++ backend/controller/anagram/anagramController.ts
@@ -2,8 +2,9 @@ import 'dotenv/config';
 
 import { Request, Response } from 'express';
 
-import { generateAnagram } from '../../utils/anagram.lib';
-import { getFactorial, getNumOfUniqueAnagrams } from '../../utils/math.lib';
+import { generateAnagram } from '../../utils/lib/anagram.lib';
+import { getFactorial, getNumOfUniqueAnagrams } from '../../utils/lib/math.lib';
+
 
 /**
  * @function getAnagrams
